Support different return types from control and candidate
Add a new type of experiment that supports control and candidate
functions that return different types. This was necessary to compare two
different code paths that return similar, but distinct types. In this
case we require a comparator to be specified, since Objects::equals does
not really make sense if the returned types are different.